// app/types/FractionalCommunication/Header.ts
// /apps/types/Fractional/FractionalHero.ts

export interface FractionalHeroSettings {
    id: number;
    headline?: string | null;
    description?: string | null;
    button_text?: string | null;
    calendar_event_title?: string | null;
    calendar_event_details?: string | null;
    calendar_event_location?: string | null;
    calendar_event_email?: string | null;
    created_at?: string;
    updated_at?: string;
  }